+++ changelog.md
@@ -31,7 +31,7 @@
 ## v0.1.1 - bottom of 34
     - Added beginnings of the ast
 
-# Nov 17th '23, v0.1.2 middle of 56
+# Nov 17th '23, v0.1.2 middle of 58
     - Added beginnings of the parser (need to finish documenting and maybe reread)
     - Added return statements to the parser parseStatement method
     - Added an initial test to the ast, in (ast_test)
@@ -41,3 +41,9 @@
     - Added parser errors
     - Added support for checking for parser errors
     - Changed parser test to test parser errors
+
+# Nov 24th '23 v0.1.3
+
+    - Added parsing capabilities for Infix expressions
+    - Added parsing capabilities for Prefix expressions
+    - Added precedence table (order of operations)